Move to hivemind.utils, add copyright notices
