+++ hivemind/p2p/p2p_daemon_bindings/control.py
@@ -9,7 +9,8 @@ from contextlib import asynccontextmanager, closing
 from typing import AsyncIterator, Awaitable, Callable, Dict, Iterable, Optional, Sequence, Tuple
 from uuid import UUID, uuid4
 
-from hivemind.p2p.multiaddr import Multiaddr, protocols
+from hivemind.utils.multiaddr import Multiaddr
+from hivemind.utils.multiaddr import protocols
 from hivemind.p2p.p2p_daemon_bindings.datastructures import PeerID, PeerInfo, StreamInfo
 from hivemind.p2p.p2p_daemon_bindings.utils import (
     DispatchFailure,
