+++ hivemind/p2p/p2p_daemon_bindings/datastructures.py
@@ -12,7 +12,7 @@ import multihash
 from cryptography.hazmat.primitives import serialization
 from multiaddr import Multiaddr
 
-from hivemind.p2p.multiaddr import Multiaddr, protocols
+from hivemind.utils.multiaddr import Multiaddr
 from hivemind.proto import crypto_pb2, p2pd_pb2
