+++ hivemind/utils/multiaddr/codecs/domain.py
@@ -1,0 +1,41 @@
+# This code is originally taken from https://github.com/multiformats/py-multiaddr
+#
+# The MIT License (MIT)
+#
+# Copyright (c) 2014-2015 Steven Buss
+# Copyright (c) 2019-2020 Alexander Schlarb
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+import idna
+
+from . import LENGTH_PREFIXED_VAR_SIZE
+
+SIZE = LENGTH_PREFIXED_VAR_SIZE
+IS_PATH = False
+
+
+def to_bytes(proto, string):
+    return idna.uts46_remap(string).encode("utf-8")
+
+
+def to_string(proto, buf):
+    string = buf.decode("utf-8")
+    for label in string.split("."):
+        idna.check_label(label)
+    return string
