+++ tests/test_dht.py
@@ -6,7 +6,7 @@ import time
 import pytest
 
 import hivemind
-from hivemind.p2p.multiaddr import Multiaddr
+from hivemind.utils.multiaddr import Multiaddr
 
 from test_utils.dht_swarms import launch_dht_instances
 from test_utils.networking import get_free_port
