+++ tests/test_p2p_daemon_bindings.py
@@ -5,7 +5,8 @@ from contextlib import AsyncExitStack
 import pytest
 from google.protobuf.message import EncodeError
 
-from hivemind.p2p.multiaddr import Multiaddr, protocols
+from hivemind.utils.multiaddr import Multiaddr
+from hivemind.utils.multiaddr import protocols
 from hivemind.p2p.p2p_daemon_bindings.control import ControlClient, DaemonConnector, parse_conn_protocol
 from hivemind.p2p.p2p_daemon_bindings.datastructures import PeerID, PeerInfo, StreamInfo
 from hivemind.p2p.p2p_daemon_bindings.utils import (
@@ -18,7 +19,7 @@ from hivemind.p2p.p2p_daemon_bindings.utils import (
 )
 from hivemind.proto import p2pd_pb2 as p2pd_pb
 
-from test_utils.p2p_daemon import connect_safe, make_p2pd_pair_unix, try_until_success
+from test_utils.p2p_daemon import connect_safe, make_p2pd_pair_unix
 
 
 def test_raise_if_failed_raises():
