+++ tests/test_utils/dht_swarms.py
@@ -8,7 +8,7 @@ from typing import Dict, List, Tuple
 from hivemind.dht import DHT
 from hivemind.dht.node import DHTID, DHTNode
 from hivemind.p2p import PeerID
-from hivemind.p2p.multiaddr import Multiaddr
+from hivemind.utils.multiaddr import Multiaddr
 
 
 def run_node(initial_peers: List[Multiaddr], info_queue: mp.Queue, **kwargs):
